+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+ - Adding scripts for managing response on quarantine queue
 
 ### 2.0.0 2018-06-14
  - OPSWAT anti-virus integration